requirements.txt
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
black==22.10.0
boto3==1.26.3
botocore==1.29.3
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.3
dacite==1.6.0
daphne==4.0.0
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.13
django-appconf==1.0.5
django-bootstrap5==22.1
django-extensions==3.2.1
django-imagekit==4.1.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-storages==1.13.1
django-widget-tweaks==1.4.12
executing==1.1.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
matplotlib-inline==0.1.6
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.10.1
phonenumbers==8.13.0
phonenumberslite==8.13.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
pokemontcgsdk==3.4.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
traitlets==5.4.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.10
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.2
